Task01/app/requirements.txt
absl-py==2.2.1
albucore==0.0.24
albumentations==2.0.7
alembic==1.15.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bcrypt==3.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
Cython==3.0.12
dbus-python==1.2.18
debugpy==1.8.13
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
duplicity==0.8.21
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
fasteners==0.14.1
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
folder-color==0.2.0
fonttools==4.56.0
fqdn==1.5.1
frozendict==2.4.6
fsspec==2025.3.0
future==0.18.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-genai==1.12.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.30.2
icoextract==0.1.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.1
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==8.10.0
mplcursors==0.6
mpmath==1.3.0
msgpack==1.0.3
multitasking==0.0.11
namex==0.0.8
narwhals==1.33.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
notebook_shim==0.2.4
numba==0.61.0
numpy==2.1.3
oauthlib==3.2.0
olefile==0.46
openai==1.76.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
patsy==1.0.1
pdf2image==1.17.0
peewee==3.17.9
pefile==2021.9.3
peopledatalabs==5.0.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
playwright==1.51.0
plotly==6.0.1
pmdarima==2.0.4
pooch==1.8.2
prometheus_client==0.22.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==5.9.0
psycopg==3.2.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
pyelftools==0.27
pygame==2.6.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.25.4
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pytesseract==0.3.13
python-apt===2.4.0-ubuntu4-zorin1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-json-logger==3.3.0
pytz==2025.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
river==0.22.0
rpds-py==0.24.0
rsa==4.9.1
rubber==1.6.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.31.0
Send2Trash==1.8.3
simsimd==6.2.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
streamlit==1.44.1
stringzilla==3.12.5
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
triton==3.2.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
uvicorn==0.34.0
virtualenv==20.29.3
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
yfinance==0.2.55
zipp==1.0.0
zorin-appearance==5.3
zorin-connect==1.2
zorin-exec-guard==1.5
